refactoring: moved privacyIndicators tests to widgets tests
